requirements.txt
# Core dependencies (in use)
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.46
PyJWT==2.6.0
requests==2.28.0
markupsafe==2.1.1

# --- Vulnerable deps NOT used by the application ---
# These are leftover from earlier features or were added "just in case"
# and never cleaned up. Dependabot will flag them, but they pose no
# actual risk since no code imports them.

PyYAML==5.3             # CVE-2020-14343 (arbitrary code exec) — not imported anywhere
Pillow==8.0.0           # Multiple CVEs (OOB read, DoS) — no image processing in app
paramiko==2.7.1         # CVE-2023-48795 (Terrapin) — no SSH functionality
lxml==4.6.2             # CVE-2021-28957 (XSS in clean) — no XML parsing in app
cryptography==3.3.2     # CVE-2023-23931 — only pulled transitively, not direct use
Jinja2==3.1.2           # Pinned but autoescaping is on; older version has edge cases
numpy==1.21.0           # CVE-2021-33430 — imported nowhere, leftover from prototype
